Move audio_analysis ui to own file
